move reducer, add/organize types
