--- src/router/util/types.js
@@ -1,2 +1,0 @@
-// @flow
-export type RouterTo = string | LocationShape | URL;
